# app-service/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - user-rest-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    put:
      tags:
      - user-rest-controller
      operationId: replaceUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
...
  /api/v1/products/{id}:
    get:
      tags:
      - product-rest-controller
      operationId: getProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      tags:
      - product-rest-controller
      operationId: replaceProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - product-rest-controller
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/images/{id}:
    get:
      tags:
      - image-rest-controller
      operationId: getImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDTO"
...
components:
  schemas:
    ImageBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    OrderBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        product:
          $ref: "#/components/schemas/ProductDTO"
        totalPrice:
          type: number
          format: double
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
    ProductBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        price:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        category:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        image:
          $ref: "#/components/schemas/ImageBasicDTO"
        isSold:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          type: string
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surnames:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        surnames:
          type: string
        email:
          type: string
        image:
          $ref: "#/components/schemas/ImageBasicDTO"
        location:
          $ref: "#/components/schemas/LocationDTO"
        createdAt:
          type: string
          format: date
        roles:
          type: array
          items:
            type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductBasicDTO"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderBasicDTO"
...
    ImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        imageFile:
          type: object
          properties:
            binaryStream: {}
...