# app-service/src/main/java/es/mqm/webapp/controller/ProductController.java
package es.mqm.webapp.controller;

import java.io.IOException;
import java.util.List;
import java.util.Map;
import java.util.Optional;

import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ResponseStatusException;
import org.springframework.web.servlet.mvc.support.RedirectAttributes;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Sort;
import org.springframework.http.HttpStatus;
import org.springframework.security.access.prepost.PreAuthorize;

import es.mqm.webapp.model.User;
import es.mqm.webapp.service.UserService;
import es.mqm.webapp.model.Product;
import es.mqm.webapp.service.ProductService;
import es.mqm.webapp.model.Review;
import es.mqm.webapp.service.ReviewService;
import es.mqm.webapp.service.ReviewUtils;
import es.mqm.webapp.model.ExtendedProduct;
import es.mqm.webapp.model.Image;
import es.mqm.webapp.service.ImageService;
import es.mqm.webapp.service.OrderService;
import org.springframework.web.bind.annotation.RequestParam;

@Controller
public class ProductController {
    @Autowired
    private UserService userService;
    @Autowired
    private ProductService productService;
    @Autowired
    private ReviewService reviewService;
    @Autowired
    private OrderService orderService;
    @Autowired
    private ImageService imageService;

    private static final int PAGE_SIZE = 3;

    @GetMapping("/product/{id}")
    public String showProductDetails(@PathVariable("id") int id, Model model,@RequestParam(value = "pageReview", defaultValue = "0") int pageReview) {
        if (pageReview < 0) {
            pageReview = 0;
        }

        User currentUser = (User) model.getAttribute("currentUser");
        ExtendedProduct extproduct = productService.findByIdWithDistance(id, currentUser).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Producto no encontrado"));
        Product product = extproduct.getP();
        if (product.getImage() != null) {
            model.addAttribute("imageUrl", product.getImage().getId());
        } else {
            model.addAttribute("imageUrl", "product-400x600.png");
        }
        if (product.getUser() != null && product.getUser().getImage() != null) {
            model.addAttribute("sellerImageUrl", product.getUser().getImage().getId());
        } else {
            model.addAttribute("sellerImageUrl", "usuario anonimo.jpg");
        }
        model.addAttribute("product", product);
        model.addAttribute("distance", extproduct.getDistance());
        Boolean isUser = currentUser != null && product.getUser() != null && product.getUser().getId() == currentUser.getId();
        model.addAttribute("isUser", isUser);
        Boolean isSold = product.getIsSold();
        model.addAttribute("isSold", isSold);

        List<Review> allSellerReviews = reviewService.findByUserDest(product.getUser().getId());
        double sellerAverage = allSellerReviews.stream().mapToDouble(Review::getRating).average().orElse(0.0);
        model.addAttribute("sellerAverage", Math.round(sellerAverage * 100.0) / 100.0);

        PageRequest reviewPageRequest = PageRequest.of(pageReview, PAGE_SIZE, Sort.by(Sort.Direction.DESC, "id"));
        Page<Review> reviewPage = reviewService.findByProductUserId(product.getUser().getId(), reviewPageRequest);
        if (pageReview >= reviewPage.getTotalPages() && reviewPage.getTotalPages() > 0) {
            pageReview = reviewPage.getTotalPages() - 1;
            reviewPageRequest = PageRequest.of(pageReview, PAGE_SIZE, Sort.by(Sort.Direction.DESC, "id"));
            reviewPage = reviewService.findByProductUserId(product.getUser().getId(), reviewPageRequest);
        }

        boolean hasNextPage = reviewPage.hasNext();
        boolean hasPreviousPage = reviewPage.hasPrevious();
        int totalPagesReview = reviewPage.getTotalPages() == 0 ? 1 : reviewPage.getTotalPages();

        model.addAttribute("hasPrevPage", hasPreviousPage);
        model.addAttribute("hasNextPage", hasNextPage);
        model.addAttribute("next", hasNextPage ? pageReview + 1 : pageReview);
        model.addAttribute("previous", hasPreviousPage ? pageReview - 1 : pageReview);
        model.addAttribute("currentPageReview", pageReview + 1);
        model.addAttribute("totalPagesReview", totalPagesReview);
        model.addAttribute("reviews", reviewPage.getContent());
        model.addAttribute("showPaginationReview", totalPagesReview > 1);

        List<Review> reviews = reviewPage.getContent();
        List<Map<String, Object>> reviewsVm = ReviewUtils.getReviewsVm(reviewPage, currentUser);
        model.addAttribute("reviewsVm", reviewsVm);
        if(reviews.isEmpty()){
            model.addAttribute("empty", true);
        }else{
            model.addAttribute("empty", false);
        }
        model.addAttribute("cssfile", "product");
        return "product";
    }

    @PostMapping("/newProduct")
    public String uploadProduct(
            Model model,
            @RequestParam String name,
            @RequestParam String description,
            @RequestParam double price,
            @RequestParam(required = false) String category,
            @RequestParam(required = false) MultipartFile image
    ) throws IOException {

        User user = userService.findById(((User) model.getAttribute("currentUser")).getId()).orElse(null);
        if (user == null) {
            return "redirect:/error";
        }
        Product product = new Product();
        product.setName(name);
        product.setDescription(description);
        product.setPrice(price);
        product.setCategory(category != null && !category.isBlank() ? category : "informatica");
        product.setUser(user);

        if (image != null && !image.isEmpty()) {
            Image im = imageService.createImage(image);
            product.setImage(im);
        }

        productService.save(product);
        return "redirect:/product/" + product.getId();
    }

    @GetMapping("/sell_product")
    public String showSellProductPage(Model model) {
        model.addAttribute("cssfile", "sell_product");
        return "sell_product";
    }

    @PreAuthorize("@productService.isOwnerOrAdmin(#id, authentication)")
    @GetMapping("/modify_product/{id}")
    public String showModifyProductPage(Model model, @PathVariable int id) {

        Optional<Product> product = productService.findById(id);
        if(product.isPresent()){
            model.addAttribute("product", product.get());
        }else{
            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Failed to load default product image");
        }
        model.addAttribute("cssfile", "sell_product");

        return "modify_product";
    }

    @PreAuthorize("@productService.isOwnerOrAdmin(#id, authentication)")
    @PostMapping("/modify_product")
    public String modifyProduct(Model model, @RequestParam int id, @RequestParam String name,
            @RequestParam String description, @RequestParam double price,
            @RequestParam(required=false) String category,
            @RequestParam(required = false) MultipartFile image, RedirectAttributes redirAttr)
            throws ResponseStatusException, IOException{
        Product product = productService.findById(id).orElse(null);
        if (product == null) {
            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Producto no encontrado");
        }
        product.setName(name);
        product.setDescription(description);
        product.setPrice(price);
        if (category != null && !category.isBlank()) {
            product.setCategory(category);
        }
        if (image != null && !image.isEmpty()) {
            Image im = imageService.createImage(image);
            productService.addImageToProduct(id, im);
        }
        productService.save(product);
        redirAttr.addFlashAttribute("toastMessage", "Producto modificado correctamente");
        return "redirect:/product/" + id;
    }

    @PreAuthorize("@productService.isOwnerOrAdmin(#id, authentication)")
    @PostMapping("delete_product_user/{id}")
    public String deleteProduct(@PathVariable int id, RedirectAttributes redirAttr) {
        int userId = productService.findById(id).map(p -> p.getUser().getId()).orElse(-1);
        if(userId == -1){
            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se ha encontrado el dueño del producto");
        }
        productService.deleteById(id);
        redirAttr.addFlashAttribute("toastMessage", "Producto eliminado correctamente");
        return "redirect:/user_profile/" + userId;
    }
}