# app-service/src/main/java/es/mqm/webapp/dto/UserDTO.java
package es.mqm.webapp.dto;

import java.time.LocalDate;
import java.util.List;

public record UserDTO(
        Integer id,
        String name,
        String surnames,
        String email,
        // String password,
        ImageBasicDTO image,
        LocationDTO location,
        LocalDate createdAt,
        List<String> roles,
        List<ProductBasicDTO> products,
        List<OrderBasicDTO> orders) {
}